requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
WTForms==3.1.1
Werkzeug==3.0.1
openai==1.60.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
python-dotenv==1.0.1
email-validator==2.1.0

